fix crash when there is a transaction with no input for bitcoin
